fix: 공시사항 발행 리팩토링
