+++ backend/src/main/java/kr/co/nogibackend/domain/admin/dto/request/NotionCreateNoticeRequest.java
@@ -4,10 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.List;
 import kr.co.nogibackend.domain.notion.dto.info.NotionBlockInfo;
 import kr.co.nogibackend.domain.notion.dto.property.NotionEmojiProperty;
-import kr.co.nogibackend.domain.notion.dto.property.NotionEmojiProperty.EMOJI_TYPE;
 import kr.co.nogibackend.domain.notion.dto.property.NotionNogiProperties;
 import kr.co.nogibackend.domain.notion.dto.property.NotionParentProperty;
-import kr.co.nogibackend.domain.notion.dto.property.NotionParentProperty.PARENT_TYPE;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record NotionCreateNoticeRequest(
@@ -16,18 +14,5 @@ public record NotionCreateNoticeRequest(
     NotionNogiProperties properties,
     List<NotionBlockInfo> children
 ) {
-
-  public static NotionCreateNoticeRequest ofNotice(
-      String databaseId,
-      String title,
-      List<NotionBlockInfo> content
-  ) {
-    return new NotionCreateNoticeRequest(
-        NotionParentProperty.buildParent(PARENT_TYPE.DATABASE, databaseId),
-        NotionEmojiProperty.buildEmoji(EMOJI_TYPE.EMOJI, "\uD83D\uDCE2"),
-        NotionNogiProperties.buildCreateNotice(title),
-        content
-    );
-  }
 
 }
