+++ backend/src/main/java/kr/co/nogibackend/domain/notice/entity/Notice.java
@@ -9,9 +9,15 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.List;
 import kr.co.nogibackend.domain.BaseEntity;
+import kr.co.nogibackend.domain.admin.dto.request.NotionCreateNoticeRequest;
 import kr.co.nogibackend.domain.notion.dto.content.NotionCalloutContent;
 import kr.co.nogibackend.domain.notion.dto.content.NotionRichTextContent;
 import kr.co.nogibackend.domain.notion.dto.info.NotionBlockInfo;
+import kr.co.nogibackend.domain.notion.dto.property.NotionEmojiProperty;
+import kr.co.nogibackend.domain.notion.dto.property.NotionEmojiProperty.EMOJI_TYPE;
+import kr.co.nogibackend.domain.notion.dto.property.NotionNogiProperties;
+import kr.co.nogibackend.domain.notion.dto.property.NotionParentProperty;
+import kr.co.nogibackend.domain.notion.dto.property.NotionParentProperty.PARENT_TYPE;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,7 +28,7 @@ import org.hibernate.envers.Audited;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Table(
-		name = "tb_notice"
+    name = "tb_notice"
 )
 @Getter
 @Entity
@@ -34,40 +40,49 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @SQLDelete(sql = "UPDATE tb_notice SET deleted = true, deleted_on = CURRENT_TIMESTAMP WHERE id = ?")
 public class Notice extends BaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-	@Column(nullable = false, length = 1000)
-	private String title;
+  @Column(nullable = false, length = 1000)
+  private String title;
 
-	@Column(nullable = false)
-	private String url;
+  @Column(nullable = false)
+  private String url;
 
-	@Column(nullable = false, columnDefinition = "LONGTEXT")
-	private String content;
+  @Column(nullable = false, columnDefinition = "LONGTEXT")
+  private String content;
 
-	@OneToMany(mappedBy = "notice")
-	private List<NoticeUser> noticeUsers;
+  @OneToMany(mappedBy = "notice")
+  private List<NoticeUser> noticeUsers;
 
-	public List<NotionBlockInfo> buildContentCallOutBlock() {
-		NotionRichTextContent linkText =
-				NotionRichTextContent.buildLinkText(this.title + " \uD83D\uDC48 공지 확인하기", this.url);
-		List<NotionRichTextContent> richTexts = List.of(linkText);
+  public NotionCreateNoticeRequest buildNewPublishToNotion(String databaseId) {
+    return new NotionCreateNoticeRequest(
+        NotionParentProperty.buildParent(PARENT_TYPE.DATABASE, databaseId),
+        NotionEmojiProperty.buildEmoji(EMOJI_TYPE.EMOJI, "\uD83D\uDCE2"),
+        NotionNogiProperties.buildNewNotice(this.title),
+        this.buildContentCallOutBlock()
+    );
+  }
 
-		return
-				List.of(
-						NotionBlockInfo
-								.builder()
-								.object("block")
-								.type(NotionBlockInfo.CALL_OUT)
-								.callout(NotionCalloutContent.buildEmoji(richTexts, "\uD83D\uDCA1"))
-								.build()
-				);
-	}
+  public List<NotionBlockInfo> buildContentCallOutBlock() {
+    NotionRichTextContent linkText =
+        NotionRichTextContent.buildLinkText(this.title + " \uD83D\uDC48 공지 확인하기", this.url);
+    List<NotionRichTextContent> richTexts = List.of(linkText);
 
-	public void updateUrl() {
-		this.url = this.url + "/" + this.id;
-	}
+    return
+        List.of(
+            NotionBlockInfo
+                .builder()
+                .object("block")
+                .type(NotionBlockInfo.CALL_OUT)
+                .callout(NotionCalloutContent.buildEmoji(richTexts, "\uD83D\uDCA1"))
+                .build()
+        );
+  }
+
+  public void updateUrl() {
+    this.url = this.url + "/" + this.id;
+  }
 
 }
