+++ backend/src/main/java/kr/co/nogibackend/interfaces/notice/NoticePublishController.java
@@ -1,6 +1,6 @@
 package kr.co.nogibackend.interfaces.notice;
 
-import kr.co.nogibackend.domain.notice.service.NoticePublishService;
+import kr.co.nogibackend.application.notice.NoticePublishFacade;
 import kr.co.nogibackend.interfaces.notice.request.NoticePublishRequest;
 import kr.co.nogibackend.response.service.Response;
 import lombok.RequiredArgsConstructor;
@@ -16,17 +16,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class NoticePublishController {
 
-  private final NoticePublishService noticePublishService;
+  private final NoticePublishFacade noticePublishFacade;
 
   @PostMapping("publish")
   public ResponseEntity<?> publish(NoticePublishRequest request) {
-    return Response.success(noticePublishService.publish(request.toCommand()));
+    return Response.success(noticePublishFacade.publish(request.toCommand()));
   }
 
   // todo: 작업필요
-  @PostMapping("re-publish")
-  public ResponseEntity<?> rePublish(NoticePublishRequest request) {
-    return Response.success(noticePublishService.publish(request.toCommand()));
-  }
+//  @PostMapping("re-publish")
+//  public ResponseEntity<?> rePublish(NoticePublishRequest request) {
+//    return Response.success(noticePublishService.publish(request.toCommand()));
+//  }
 
 }
